+++ packages/dashboard/src/components/iot-dashboard/iot-dashboard-widget/iot-dashboard-widget.tsx
@@ -26,8 +26,8 @@ export class IotDashboardWidget {
           height: `${cellSize * height}px`,
         }}
       >
-        <div class={`widget ${this.isSelected && 'widget-selected'}`} id={id}>
-          <div class="widget-content">{this.widget.widget}</div>
+        <div className={`widget ${this.isSelected && 'widget-selected'}`} id={id}>
+          <div className="widget-content">{this.widget.widget}</div>
         </div>
       </Host>
     );