+++ packages/dashboard/src/dashboard-actions/createWidget.ts
@@ -1,0 +1,12 @@
+import { concat } from 'lodash';
+import { CreateAction, DashboardConfiguration, Widget } from '../types';
+
+export const createWidget = ({
+  dashboardConfiguration,
+  widgets,
+}: {
+  dashboardConfiguration: DashboardConfiguration;
+  widgets: Widget[];
+}): DashboardConfiguration => {
+  return concat(dashboardConfiguration, widgets);
+};