+++ packages/dashboard/src/types.ts
@@ -62,4 +62,107 @@ export const onResizeAction = (payload: ResizeAction['payload']): ResizeAction =
 
 export type ResizeActionInput = ResizeAction['payload'];
 
-export type DashboardAction = MoveAction | ResizeAction;
+export const DELETE = 'DELETE';
+
+export interface DeleteAction extends Action<'DELETE'> {
+  type: typeof DELETE;
+  payload: {
+    widgetIds: string[];
+    widgets: Widget[];
+  };
+}
+export const onDeleteAction = (payload: DeleteAction['payload']): DeleteAction => ({
+  type: DELETE,
+  payload,
+});
+
+export type DeleteActionInput = DeleteAction['payload'];
+
+export const PASTE = 'PASTE';
+
+export interface PasteAction extends Action<'PASTE'> {
+  type: typeof PASTE;
+  payload: {
+    copyGroup: Widget[];
+    numTimesCopyGroupHasBeenPasted: number;
+  };
+}
+export const onPasteAction = (payload: PasteAction['payload']): PasteAction => ({
+  type: PASTE,
+  payload,
+});
+
+export type PasteActionInput = PasteAction['payload'];
+
+export const CREATE = 'CREATE';
+
+export interface CreateAction extends Action<'CREATE'> {
+  type: typeof CREATE;
+  payload: {
+    dashboardConfiguration: DashboardConfiguration;
+    widgets: Widget[];
+  };
+}
+
+export const onCreateAction = (payload: CreateAction['payload']): CreateAction => ({
+  type: CREATE,
+  payload,
+});
+
+export type CreateActionInput = CreateAction['payload'];
+
+export const REDO = 'REDO';
+
+export interface RedoAction extends Action<'REDO'> {
+  type: typeof REDO;
+  payload: {
+    redoAction: DashboardAction;
+  };
+}
+
+export type RedoActionInput = RedoAction['payload'];
+
+export const onRedoAction = (payload: RedoActionInput): RedoAction => ({
+  type: REDO,
+  payload,
+});
+
+export const UNDO = 'UNDO';
+
+export interface UndoAction extends Action<'UNDO'> {
+  type: typeof UNDO;
+  payload: {
+    undoAction: DashboardAction;
+  };
+}
+
+export type UndoActionInput = UndoAction['payload'];
+
+export const onUndoAction = (payload: UndoActionInput): UndoAction => ({
+  type: UNDO,
+  payload,
+});
+
+export type DashboardAction =
+  | MoveAction
+  | ResizeAction
+  | DeleteAction
+  | PasteAction
+  | CreateAction
+  | UndoAction
+  | VoidAction
+  | RedoAction;
+
+export const VOID = 'VOID';
+
+export interface VoidAction extends Action<'VOID'> {
+  type: typeof VOID;
+  payload: {};
+}
+
+export const onVoidAction = (): VoidAction => ({
+  type: VOID,
+  payload: {},
+});
+
+export type UndoQueue = DashboardAction[];